+++ controllers/Reception.js
@@ -52,34 +52,34 @@ exports.addReception = async (req, res) => {
           .input("acknowledgement", Acknowledgement)
           .input("departmentVisit", DepartmentVisit)
           .execute("dbo.Visits_Insert");
+      }
+      // If there is a record, check if the user has signed in or out
+      else if(latestVisitRecord.recordset[0].signInDate && latestVisitRecord.recordset[0].signOutDate ){
+        await pool
+        .request()
+        .input("firstName", FirstName)
+        .input("lastName", LastName)
+        .input("phoneNumber", PhoneNumber)
+        .input("signInDate", SignInOutTime)
+        .input("homeAreas", JSON.stringify(HomeAreas))
+        .input("scheduledVisit", ScheduledVisit)
+        .input("purpose", Purpose)
+        .input("residentName", ResidentName)
+        .input("firstVisit", FirstVisit)
+        .input("sicknessSymptom", SicknessSymptom)
+        .input("acknowledgement", Acknowledgement)
+        .input("departmentVisit", DepartmentVisit)
+        .execute("dbo.Visits_Insert");
       } else {
         if (
           latestVisitRecord.recordset[0].signInDate &&
           !latestVisitRecord.recordset[0].signOutDate
         ) {
-          return res
-            .status(200)
-            .json({
-              success: false,
-              error:
-                "You have not signed out yet. Please sign out before continue",
-            });
-        } else {
-          await pool
-            .request()
-            .input("firstName", FirstName)
-            .input("lastName", LastName)
-            .input("phoneNumber", PhoneNumber)
-            .input("signInDate", SignInOutTime)
-            .input("homeAreas", JSON.stringify(HomeAreas))
-            .input("scheduledVisit", ScheduledVisit)
-            .input("purpose", Purpose)
-            .input("residentName", ResidentName)
-            .input("firstVisit", FirstVisit)
-            .input("sicknessSymptom", SicknessSymptom)
-            .input("acknowledgement", Acknowledgement)
-            .input("departmentVisit", DepartmentVisit)
-            .execute("dbo.Visits_Insert");
+          return res.status(200).json({
+            success: false,
+            error:
+              "You have not signed out yet. Please sign out before continue",
+          });
         }
       }
     }
@@ -117,10 +117,6 @@ exports.updateReception = async (req, res) => {
   try {
     console.log("test my request body", req.body, req.params.receptionID);
 
-    // Update from the portal
-    // if(req.params.receptionID){
-    //   return
-    // }
     // Update from form data
     if (req.body) {
       const { FirstName, LastName, PhoneNumber, InOut } = req.body;
@@ -140,26 +136,34 @@ exports.updateReception = async (req, res) => {
         .input("latestVisit", 1)
         .execute("dbo.Visits_Load");
 
-      console.log("test field", latestVisitRecord.recordset);
       // First time visit then insert new record
       if (latestVisitRecord.recordset.length === 0) {
-        return res
-          .status(200)
-          .json({
-            success: false,
-            error:
-              "Your name was not found! Please check the spelling of your name.",
-          });
+        return res.status(200).json({
+          success: false,
+          error:
+            "Your name was not found! Please check the spelling of your name.",
+        });
       } else {
-        await pool
-          .request()
-          .input("visitID", latestVisitRecord.recordset[0].visitID)
-          .input("signOutDate", SignInOutTime)
-          .execute("dbo.Visits_Update");
+        if (
+          latestVisitRecord.recordset[0].signInDate &&
+          latestVisitRecord.recordset[0].signOutDate
+        ) {
+          return res.status(200).json({
+            success: false,
+            error: "You have not signed in yet. Please sign in before continue",
+          });
+        } else {
+          await pool
+            .request()
+            .input("visitID", latestVisitRecord.recordset[0].visitID)
+            .input("signOutDate", SignInOutTime)
+            .execute("dbo.Visits_Update");
+          res.status(200).json({ success: true, error: null });
+        }
       }
+    } else {
+      res.status(200).json({ success: false, error: "Bad Request" });
     }
-
-    res.status(200).json({ success: true });
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: "Server Error" });