+++ utils/storage.js
@@ -99,8 +99,9 @@ async function retrieveImage(subloc, id) {
 
   // Local
   // const url = `${process.env.BACKEND_BASE_URL}/assets/${subloc}/${fileName}`;
+  // const url = `https://b9dk2wds-5001.use.devtunnels.ms/assets/${subloc}/${fileName}`;
 
-  let img = fs.readFileSync(imagePath, { encoding: "base64" });
+  // let img = fs.readFileSync(imagePath, { encoding: "base64" });
 
   return url;
 }
@@ -125,8 +126,9 @@ async function retrieveAudio(subloc, id) {
   const url = `https://trinityvillagedev.azurewebsites.net/assets/${subloc}/${fileName}`;
 
   // const url = `${process.env.BACKEND_BASE_URL}/assets/${subloc}/${fileName}`;
+  // const url = `https://b9dk2wds-5001.use.devtunnels.ms/assets/${subloc}/${fileName}`;
 
-  let audio = fs.readFileSync(audioPath, { encoding: "base64" });
+  // let audio = fs.readFileSync(audioPath, { encoding: "base64" });
 
   return url;
 }